# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Think Tank CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CrmUser:
      type: object
      required: [id, email, role, status, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
        status:
          type: string
...