# lib/api-zod/src/generated/types/crmUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Think Tank CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CrmUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: string;
  status: string;
  createdAt: string;
}